+++ src/use-google-classroom.ts
@@ -5,6 +5,7 @@ const SCOPES = [
   'https://www.googleapis.com/auth/classroom.courses.readonly',
   'https://www.googleapis.com/auth/classroom.coursework.me.readonly',
   'https://www.googleapis.com/auth/classroom.courseworkmaterials.readonly',
+  'https://www.googleapis.com/auth/classroom.announcements.readonly',
   'https://www.googleapis.com/auth/drive.readonly',
 ].join(' ')